solution at live 5

+++ src/main/java/com/cubixedu/hr/sample/service/HolidayRequestService.java
@@ -9,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
@@ -17,6 +19,7 @@ import com.cubixedu.hr.sample.dto.HolidayRequestFilterDto;
 import com.cubixedu.hr.sample.model.Employee;
 import com.cubixedu.hr.sample.model.HolidayRequest;
 import com.cubixedu.hr.sample.repository.HolidayRequestRepository;
+import com.cubixedu.hr.sample.security.HrUser;
 
 
 @Service
@@ -73,6 +76,11 @@ public class HolidayRequestService {
 	@Transactional
 	public HolidayRequest approveHolidayRequest(long id, long approverId, boolean status) {
 		HolidayRequest holidayRequest = holidayRequestRepository.findById(id).get();
+		
+		Employee currentUser = getCurrentHrUser().getEmployee();
+		if(!holidayRequest.getEmployee().getManager().getEmployeeId().equals(currentUser.getEmployeeId()))
+			throw new AccessDeniedException("Only the manager of the user can approve the request");
+		
 		holidayRequest.setApprover(employeeService.findById(approverId).get());
 		holidayRequest.setApproved(status);
 		holidayRequest.setApprovedAt(LocalDateTime.now());
@@ -83,7 +91,7 @@ public class HolidayRequestService {
 	public HolidayRequest modifyHolidayRequest(long id, HolidayRequest newHolidayRequest) {
 		HolidayRequest holidayRequest = holidayRequestRepository.findById(id).get();
 		if (holidayRequest.getApproved() != null)
-			throw new IllegalStateException();
+			throw new IllegalStateException();		
 		holidayRequest.setEndDate(newHolidayRequest.getEndDate());
 		holidayRequest.setStartDate(newHolidayRequest.getStartDate());
 		holidayRequest.setCreatedAt(LocalDateTime.now());
@@ -95,8 +103,17 @@ public class HolidayRequestService {
 		HolidayRequest holidayRequest = holidayRequestRepository.findById(id).get();
 		if (holidayRequest.getApproved() != null)
 			throw new IllegalArgumentException();
-		holidayRequest.getEmployee().getHolidayRequests().remove(holidayRequest);
+				
+		Employee employee = holidayRequest.getEmployee();
+		if(!employee.getEmployeeId().equals(getCurrentHrUser().getEmployee().getEmployeeId()))
+			throw new AccessDeniedException("Trying to delete other user's holiday request");
+			
+		employee.getHolidayRequests().remove(holidayRequest);
 		holidayRequestRepository.deleteById(id);
+	}
+
+	private HrUser getCurrentHrUser() {
+		return (HrUser)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 	}
 
 }
